core: Remove __DATE_YEAR__
This has not been in use for years and it's time to go away---it is the only
thing in TAME that causes nondeterminism, at least that I'm immediately
aware of.  Perhaps I'll find something else while reimplementing TAME in
TAMER.

_This does not remove the compiler code to produce this._  If something
still needs `__DATE_YEAR__` (because it's really old), it can define this
value itself, and still utilize it until TAMER (which will not include it).

DEV-7145