tame: bootstrap: Stop building after bootstrap
This was a relic of the old bootstrap system, where bootstrapping was in the
context of a parent project that utilized TAME (and so TAME needed to be
built).  But that doesn't make sense in the context of TAME itself, and what
_part_ of TAME should be built should be controlled by the project utilizing
it.

This is especially important now that TAMER builds are getting much longer
with the introduction of NIR and its parser-generator.

DEV-7145